未改url
只差和其他部分对接以及编辑资料

+++ app/src/main/java/com/example/yanghan/gravity/data/model/Reply.java
@@ -3,9 +3,9 @@ package com.example.yanghan.gravity.data.model;
 import java.util.Date;
 
 public class Reply {
-    public float replyid;
+    public int replyid;
 
-    public float postid;
+    public int postid;
 
     public String posterid="";
     public String posterName="";
